+++ rs/consensus/idkg/src/payload_builder.rs
@@ -10,7 +10,7 @@ use ic_consensus_utils::pool_reader::PoolReader;
 use ic_crypto::retrieve_mega_public_key_from_registry;
 use ic_interfaces::idkg::IDkgPool;
 use ic_interfaces_registry::RegistryClient;
-use ic_interfaces_state_manager::StateManager;
+use ic_interfaces_state_manager::StateReader;
 use ic_logger::{ReplicaLogger, error, info, warn};
 use ic_registry_client_helpers::subnet::SubnetRegistry;
 use ic_registry_subnet_features::ChainKeyConfig;
@@ -350,9 +350,6 @@ fn create_summary_payload_helper(
 
     idkg_summary.idkg_transcripts.clear();
 
-    // Purge deprecated signature agreements in the idkg payload.
-    idkg_summary.signature_agreements.clear();
-
     // We purge available pre-signatures of the parent payload,
     // because they were already delivered with the previous payload.
     idkg_summary.available_pre_signatures.clear();
@@ -483,7 +480,7 @@ pub fn create_data_payload(
     registry_client: &dyn RegistryClient,
     pool_reader: &PoolReader<'_>,
     idkg_pool: Arc<RwLock<dyn IDkgPool>>,
-    state_manager: &dyn StateManager<State = ReplicatedState>,
+    state_reader: &dyn StateReader<State = ReplicatedState>,
     context: &ValidationContext,
     parent_block: &Block,
     idkg_payload_metrics: &IDkgPayloadMetrics,
@@ -535,7 +532,7 @@ pub fn create_data_payload(
         &summary_block,
         &block_reader,
         &transcript_builder,
-        state_manager,
+        state_reader,
         registry_client,
         Some(idkg_payload_metrics),
         log,
@@ -575,7 +572,7 @@ pub(crate) fn create_data_payload_helper(
     summary_block: &Block,
     block_reader: &dyn IDkgBlockReader,
     transcript_builder: &dyn IDkgTranscriptBuilder,
-    state_manager: &dyn StateManager<State = ReplicatedState>,
+    state_reader: &dyn StateReader<State = ReplicatedState>,
     registry_client: &dyn RegistryClient,
     idkg_payload_metrics: Option<&IDkgPayloadMetrics>,
     log: &ReplicaLogger,
@@ -601,7 +598,7 @@ pub(crate) fn create_data_payload_helper(
     };
 
     let receivers = get_subnet_nodes(registry_client, next_interval_registry_version, subnet_id)?;
-    let state = state_manager.get_state_at(context.certified_height)?;
+    let state = state_reader.get_state_at(context.certified_height)?;
 
     let reshare_contexts = state.get_ref().reshare_chain_key_contexts();
     let idkg_dealings_contexts = filter_idkg_reshare_chain_key_contexts(reshare_contexts);
@@ -652,9 +649,6 @@ pub(crate) fn create_data_payload_helper_2(
     // We purge available pre-signatures of the parent payload,
     // because they were already delivered with the previous payload.
     idkg_payload.available_pre_signatures.clear();
-
-    // Purge deprecated signature agreements in the idkg payload.
-    idkg_payload.signature_agreements.clear();
 
     let new_transcripts = [
         pre_signatures::update_pre_signatures_in_creation(
@@ -1301,13 +1295,6 @@ mod tests {
             .unwrap();
             assert_eq!(result.len(), 1);
 
-            idkg_payload
-                .signature_agreements
-                .insert([2; 32], idkg::CompletedSignature::ReportedToExecution);
-            idkg_payload.signature_agreements.insert(
-                [3; 32],
-                idkg::CompletedSignature::Unreported(empty_response()),
-            );
             idkg_payload.xnet_reshare_agreements.insert(
                 create_reshare_request(key_id, 6, 6),
                 idkg::CompletedReshareRequest::ReportedToExecution,
@@ -1350,24 +1337,6 @@ mod tests {
                 Ok(())
             );
 
-            let (reported, unreported) = {
-                let mut reported = 0;
-                let mut unreported = 0;
-                for agreement in summary.signature_agreements.values() {
-                    match agreement {
-                        idkg::CompletedSignature::ReportedToExecution => {
-                            reported += 1;
-                        }
-                        idkg::CompletedSignature::Unreported(_) => {
-                            unreported += 1;
-                        }
-                    }
-                }
-                (reported, unreported)
-            };
-            assert!(!summary.signature_agreements.is_empty());
-            assert!(reported > 0);
-            assert!(unreported > 0);
             assert!(!summary.available_pre_signatures.is_empty());
             assert!(!summary.pre_signatures_in_creation.is_empty());
             assert!(!summary.idkg_transcripts.is_empty());
@@ -1409,24 +1378,9 @@ mod tests {
             assert_proposal_conversion(b);
 
             // Convert to proto format and back
-            let mut summary_proto = pb::IDkgPayload::from(&summary);
+            let summary_proto = pb::IDkgPayload::from(&summary);
             let summary_from_proto = IDkgPayload::try_from(summary_proto.clone()).unwrap();
             assert_eq!(summary, summary_from_proto);
-
-            // Check signature_agreement upgrade compatibility
-            summary_proto
-                .signature_agreements
-                .push(pb::CompletedSignature {
-                    pseudo_random_id: vec![4; 32],
-                    unreported: None,
-                });
-            let summary_from_proto = IDkgPayload::try_from(summary_proto).unwrap();
-            // Make sure the previous RequestId record can be retrieved by its pseudo_random_id.
-            assert!(
-                summary_from_proto
-                    .signature_agreements
-                    .contains_key(&[4; 32])
-            );
         })
     }