+++ rs/execution_environment/src/lib.rs
@@ -334,6 +334,7 @@ fn setup_execution_helper(
             logger.clone(),
             Arc::clone(&cycles_account_manager),
             wasm_executor,
+            config.embedders_config.create_execution_state_base_cost,
             config.embedders_config.cost_to_compile_wasm_instruction,
             config.embedders_config.dirty_page_overhead,
             config.canister_guaranteed_callback_quota,
@@ -345,7 +346,6 @@ fn setup_execution_helper(
         logger.clone(),
         metrics_registry,
         completed_execution_messages_tx,
-        Arc::clone(&state_reader),
     ));
     let ingress_history_reader = Box::new(IngressHistoryReaderImpl::new(Arc::clone(&state_reader)));