+++ rs/execution_environment/src/scheduler/test_utilities.rs
@@ -654,6 +654,7 @@ impl SchedulerTest {
             .charge_canisters_for_resource_allocation_and_usage(
                 self.state.as_mut().unwrap(),
                 subnet_size,
+                ExecutionRound::from(0),
             )
     }
 
@@ -807,13 +808,15 @@ impl Default for SchedulerTestBuilder {
         let subnet_type = SubnetType::Application;
         let scheduler_config = SubnetConfig::new(subnet_type).scheduler_config;
         let config = ic_config::execution_environment::Config::default();
+        let mut hypervisor_config = config.embedders_config;
+        hypervisor_config.create_execution_state_base_cost = NumInstructions::from(0);
         Self {
             own_subnet_id: subnet_test_id(1),
             nns_subnet_id: subnet_test_id(2),
             subnet_type,
             batch_time: UNIX_EPOCH,
             scheduler_config,
-            hypervisor_config: config.embedders_config,
+            hypervisor_config,
             initial_canister_cycles: Cycles::new(1_000_000_000_000_000_000),
             subnet_memory_capacity: config.subnet_memory_capacity.get(),
             subnet_guaranteed_response_message_memory: config
@@ -1046,6 +1049,7 @@ impl SchedulerTestBuilder {
             canister_guaranteed_callback_quota: self.canister_guaranteed_callback_quota,
             rate_limiting_of_instructions,
             rate_limiting_of_heap_delta,
+            embedders_config: self.hypervisor_config.clone(),
             ..ic_config::execution_environment::Config::default()
         };
         let wasm_executor = Arc::new(TestWasmExecutor::new(
@@ -1484,6 +1488,14 @@ impl TestWasmExecutorCore {
         let prepayment_for_response_transmission = self
             .cycles_account_manager
             .prepayment_for_response_transmission(self.subnet_size, system_state.cost_schedule());
+        // Scheduler uses `TestCall` requests which have zero payload.
+        let payload_size = NumBytes::from(0);
+        let prepayment_for_call_transmission =
+            self.cycles_account_manager.xnet_total_transmission_fee(
+                payload_size,
+                self.subnet_size,
+                system_state.cost_schedule(),
+            );
         let deadline = NO_DEADLINE;
         let callback = system_state
             .register_callback(Callback {
@@ -1492,6 +1504,7 @@ impl TestWasmExecutorCore {
                 cycles_sent: Cycles::zero(),
                 prepayment_for_response_execution,
                 prepayment_for_response_transmission,
+                prepayment_for_call_transmission,
                 on_reply: closure.clone(),
                 on_reject: closure,
                 on_cleanup: None,