+++ rs/replicated_state/tests/replicated_state.rs
@@ -1224,9 +1224,6 @@ fn split() {
     expected.metadata.ingress_history = fixture.state.metadata.ingress_history.clone();
     // And the split marker should be set.
     expected.metadata.split_from = Some(SUBNET_A);
-    // The logic preserves the full subnet schedule, even though only the priority
-    // of `CANISTER_2` will be persisted along with the canister state.
-    expected.metadata.subnet_schedule = fixture.state.metadata.subnet_schedule;
     // Otherwise, the state should be the same.
     assert_eq!(expected, state_b);